# docs/swagger.yaml
definitions:
  handlers.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  services.CartItemResponse:
    properties:
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      total:
        type: number
    type: object
  utils.CategoryRequest:
    properties:
      name:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  utils.UpdateItemRequest:
    properties:
      quantity:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/admin/categories:
    post:
      consumes:
      - application/json
      description: Добавляет новую категорию (только для администратора)
      parameters:
      - description: Название категории
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/utils.CategoryRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Category created
          schema:
            type: string
        "400":
          description: Invalid body or duplicate name
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать новую категорию
      tags:
      - Категории
  /api/admin/categories/{id}:
    delete:
      description: Удаляет категорию по ID (только для администратора)
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Category deleted
          schema:
            type: string
        "400":
          description: Удаление не удалось
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить категорию
      tags:
      - Категории
    put:
      consumes:
      - application/json
      description: Изменяет название категории (только для администратора)
      parameters:
      - description: Идентификатор категории
        in: path
        name: id
        required: true
        type: integer
      - description: Новое имя категории
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/utils.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить категорию
      tags:
      - Категории
  /api/admin/logs:
    get:
      description: Возвращает содержимое лог-файла (только для администратора)
      produces:
      - text/plain
      responses:
        "200":
          description: Содержимое лог-файла
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить лог-файл
      tags:
      - Логи
...
  /api/cart:
    get:
      description: Возвращает список товаров в корзине для owner_id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CartItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Получить содержимое корзины
      tags:
      - Корзина
    post:
      consumes:
      - application/json
      description: Добавляет товар в корзину по owner_id (user или ip)
      parameters:
      - description: ID товара и количество
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToCartRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Added to cart
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Добавить товар в корзину
      tags:
      - Корзина
  /api/cart/{product_id}:
    delete:
      description: Удаляет товар по ID из корзины owner_id
      parameters:
      - description: Идентификатор товара, который нужно удалить или обновить
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Item deleted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Удалить товар из корзины
      tags:
      - Корзина
    put:
      consumes:
      - application/json
      description: Обновляет количество указанного товара для owner_id
      parameters:
      - description: ID товара
        in: path
        name: product_id
        required: true
        type: integer
      - description: Новое количество
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/utils.UpdateItemRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Quantity updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Обновить количество товара в корзине
      tags:
      - Корзина
  /api/cart/clear:
    delete:
      description: Удаляет все товары из корзины owner_id
      produces:
      - text/plain
      responses:
        "200":
          description: Cart cleared
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Очистить корзину
      tags:
      - Корзина
  /api/categories:
    get:
      description: Возвращает все доступные категории товаров
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Получить список категорий
      tags:
      - Категории
...
swagger: "2.0"